requirements.txt
transformers
datasets
accelerate
peft
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
PyYAML
tqdm
pymupdf
distilabel
bitsandbytes
scipy
trl

